+++ hwbench/bench/test_benchmarks.py
@@ -10,7 +10,7 @@ class TestParse(tbc.TestCommon):
         # We need to patch list_module_parameters() function
         # to avoid considering the local stress-ng binary
         with patch(
-            "hwbench.engines.stressng.EngineModuleCpu.list_module_parameters"
+            "hwbench.engines.stressng_cpu.EngineModuleCpu.list_module_parameters"
         ) as p:
             print(pathlib.Path("."))
             p.return_value = (
@@ -98,7 +98,7 @@ class TestParse(tbc.TestCommon):
 
     def test_stream_short(self):
         with patch(
-            "hwbench.engines.stressng.EngineModuleCpu.list_module_parameters"
+            "hwbench.engines.stressng_cpu.EngineModuleCpu.list_module_parameters"
         ) as p:
             p.return_value = (
                 pathlib.Path("./hwbench/tests/parsing/stressngmethods/v17/stdout")
