+++ hwbench/engines/stressng.py
@@ -1,6 +1,4 @@
-import os
 import re
-import subprocess
 from collections.abc import Iterable
 from typing import NamedTuple, Callable
 
@@ -118,46 +116,12 @@ class EngineModuleVNNI(EngineModulePinnable):
         return msg
 
 
-class EngineModuleCpu(EngineModulePinnable):
-    """This class implements the EngineModuleBase for StressNG"""
-
-    def __init__(self, engine: EngineBase, engine_module_name: str):
-        super().__init__(engine, engine_module_name)
-        self.engine_module_name = engine_module_name
-        self.load_module_parameter()
-
-    def list_module_parameters(self):
-        english_env = os.environ.copy()
-        english_env["LC_ALL"] = "C"
-        cmd_out = subprocess.run(
-            [self.engine.get_binary(), "--cpu-method", "list"],
-            capture_output=True,
-            env=english_env,
-            stdin=subprocess.DEVNULL,
-        )
-        return (cmd_out.stdout or cmd_out.stderr).split(b":", 1)
-
-    def load_module_parameter(self):
-        out = self.list_module_parameters()
-        methods = out[1].decode("utf-8").split()
-        methods.remove("all")
-        for method in methods:
-            self.add_module_parameter(method)
-
-    def run_cmd(self, p: BenchmarkParameters):
-        return StressNGCPU(self, p).run_cmd()
-
-    def run(self, p: BenchmarkParameters):
-        return StressNGCPU(self, p).run()
-
-    def fully_skipped_job(self, p) -> bool:
-        return StressNGCPU(self, p).fully_skipped_job()
-
-
 class Engine(EngineBase):
     """The main stressn2 class."""
 
     def __init__(self):
+        from .stressng_cpu import EngineModuleCpu
+
         super().__init__("stressng", "stress-ng")
         self.add_module(EngineModuleCpu(self, "cpu"))
         self.add_module(EngineModuleQsort(self, "qsort"))
@@ -294,19 +258,6 @@ class StressNG(ExternalBench):
             "effective_runtime": 0,
             "skipped": True,
         }
-
-
-class StressNGCPU(StressNG):
-    """The StressNG CPU stressor."""
-
-    def run_cmd(self) -> list[str]:
-        # Let's build the command line to run the tool
-        return super().run_cmd() + [
-            "--cpu",
-            str(self.parameters.get_engine_instances_count()),
-            "--cpu-method",
-            self.parameters.get_engine_module_parameter(),
-        ]
 
 
 class StressNGQsort(StressNG):
